+++ CHANGELOG.md
@@ -41,6 +41,25 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [3.3.0](https://github.com/technway/technway-component-library/compare/root-v3.2.0...root-v3.3.0) (2025-01-08)
+
+
+### Features
+
+* add props to `tnw-contact-banner` ([4cbe247](https://github.com/technway/technway-component-library/commit/4cbe24750b96a63ffb15463e3495f8a9e32dac44))
+* **tnw-button:** implement hover effects and styling ([5e3be35](https://github.com/technway/technway-component-library/commit/5e3be358a8f5dba15f3177c04ed7aa68c2663acb))
+* **tnw-card:** implement new props and fix styling issues ([467a769](https://github.com/technway/technway-component-library/commit/467a76936f1115bd9467b089df1fe294c622ea90))
+* **tnw-image:** introduce link prop for linkable images ([8a519cc](https://github.com/technway/technway-component-library/commit/8a519cce6768349dba2d900c9f1846ab01e35359))
+* **tnw-input:** add appearanceColor prop ([794a8a1](https://github.com/technway/technway-component-library/commit/794a8a1a894b7566f529725b042d4912cf675122))
+* **tnw-navbar:** add new props for more flexibility ([0f4ad16](https://github.com/technway/technway-component-library/commit/0f4ad160f53fa0de5d4d6666d2fda83c17f8cdba))
+* **tnw-navbar:** add optional menu slot ([76a1637](https://github.com/technway/technway-component-library/commit/76a16373db5a37c4cb0bd41e09dda4c6578bc119))
+* **tnw-subscription-form:** align input appearance with theme prop ([6c36885](https://github.com/technway/technway-component-library/commit/6c36885bfbc63ae3b86493228efb1cf4403cc46f))
+
+
+### Bug Fixes
+
+* **tnw-image:** resolve borderRadius prop issues ([4a0aea2](https://github.com/technway/technway-component-library/commit/4a0aea23c5776a3c16efb6967491240be87a80b5))
+
 ## [3.2.0](https://github.com/technway/technway-component-library/compare/root-v3.1.0...root-v3.2.0) (2025-01-08)
 
 
@@ -203,7 +222,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### Features
 
 * **components:** add highlight text support to typography components ([b61d42b](https://github.com/technway/technway-component-library/commit/b61d42b709072c0d614509515ee29d780d8d8daa))
-* **components:** improve tnw-footer and tnw-newsletter-form ([0894572](https://github.com/technway/technway-component-library/commit/08945725a614f5611cdd02bb6620419d45ab4ccd))
+* **components:** improve tnw-footer and tnw-subscription-form ([0894572](https://github.com/technway/technway-component-library/commit/08945725a614f5611cdd02bb6620419d45ab4ccd))
 * **tnw-card:** add 'largerImage' prop ([4583619](https://github.com/technway/technway-component-library/commit/458361904814e3894dd8633c806e721dfa372395))
 * **tnw-copyrights-footer:** centralize content within a single &lt;tnw-text&gt; element ([3680fbf](https://github.com/technway/technway-component-library/commit/3680fbfdef6f03961ee5e574952d23355b7ea2ad))
 * **tnw-footer:** add 'copyrights' slot and improve styling with new props ([44982b0](https://github.com/technway/technway-component-library/commit/44982b0968b68bad68905cf7b5a9df145da878d9))
@@ -219,9 +238,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-* **components:** improve tnw-copyrights-footer, tnw-newsletter-form, and tnw-footer ([cbe31c9](https://github.com/technway/technway-component-library/commit/cbe31c943d5255e597ea96eb0bdd56066ac5fa1d))
+* **components:** improve tnw-copyrights-footer, tnw-subscription-form, and tnw-footer ([cbe31c9](https://github.com/technway/technway-component-library/commit/cbe31c943d5255e597ea96eb0bdd56066ac5fa1d))
 * **components:** introduce tnw-footer component and update existing footer structure ([ea30a29](https://github.com/technway/technway-component-library/commit/ea30a29a6750fbead5d71036f3e3319f6f41443b))
-* **tnw-newsletter-form:** introduce new component ([23baeb7](https://github.com/technway/technway-component-library/commit/23baeb701053070d42e5a3e327423b62a3739052))
+* **tnw-subscription-form:** introduce new component ([23baeb7](https://github.com/technway/technway-component-library/commit/23baeb701053070d42e5a3e327423b62a3739052))
 
 ## [2.6.1](https://github.com/technway/technway-component-library/compare/root-v2.6.0...root-v2.6.1) (2024-12-03)